CS5330_Part-1-main/module7/mini-project7/requirements.txt
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.3.0
six==1.16.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
tzdata==2024.2
